Creazione rotte event

+++ app/Controllers/Http/CategoryController.js
@@ -19,8 +19,10 @@ class CategoryController {
   async getById ({response, params}) {
     try {
       const categories = await Category.find(params.id)
-      return response.status(200).send(categories)
-    } catch(e) {
+        if(categories) {
+          return response.status(200).send(categories)
+        }else response.status(404).send("Categoria non trovata")
+      } catch(e) {
       return response.status(500).send({
         message: e.message
       })
@@ -30,8 +32,10 @@ class CategoryController {
   async deleteById({params, response}) {
   try {
     const categories = await Category.find(params.id)
-    await categories.delete()
-    return  response.status(200).send("Categoria Cancellata Correttamente")
+    if(categories) {
+      await categories.delete()
+      return  response.status(200).send("Categoria Cancellata Correttamente")
+    }else response.status(404).send("Categoria non trovata")
   } catch(e) {
     return response.status(500).send({
       message: e.message
@@ -43,9 +47,11 @@ async updateById ({ params, request, response }) {
   try {
     const nome = request.only(['nome'])
     const categories = await Category.find(params.id)
-    categories.merge({...nome})
-    await categories.save()
-    return  response.status(200).send("Categoria Modificata Correttamente")
+    if(categories) {
+      categories.merge({...nome})
+      await categories.save()
+      return  response.status(200).send("Categoria Modificata Correttamente")
+    }else response.status(404).send("Categoria non trovata")
   } catch(e) {
     return response.status(500).send({
       message: e.message
